backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
numpy
scikit-learn
ta
python-dotenv
requests
httpx
pydantic
